requirements.txt
aiofiles==23.2.1
aiosmtplib==2.0.2
alembic==1.12.1
asgiref==3.7.2
asyncmy==0.2.8
bcrypt==4.0.1
cryptography==41.0.4
email-validator==2.0.0
fast-captcha==0.2.1
fastapi==0.104.0
fastapi-limiter==0.1.5
fastapi-pagination==0.12.11
loguru==0.7.2
passlib==1.7.4
path==15.1.2
pre-commit==3.5.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
python-multipart==0.0.6
redis[hiredis]==4.6.0
SQLAlchemy==2.0.22
starlette==0.27.0
supervisor==4.2.5
tzdata==2023.3
uvicorn[standard]==0.23.2
wait-for-it==2.2.2
